\name{covxfrm}
\alias{covxfrm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Covariance matrix of a linear transformation}
\description{To compute the covariance matrix of \eqn{AY}, where \eqn{A}$ is a matrix and \eqn{Y} is a random vector with known covariance matrix.
}
\usage{
covxfrm(covmat, A)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{covmat}{A matrix object containing the covariance matrix of \eqn{Y}
}
  \item{A}{The matrix \eqn{A}
}
}
\value{The required covariance matrix}
\author{
Richard E. Chandler <r.chandler@ucl.ac.uk>
}
