\name{PECTSP-package}
\alias{PECTSP-package}
\alias{PECTSP}
\docType{package}
\title{\packageTitle{PECTSP}}
\description{\packageDescription{PECTSP}}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PECTSP}
\packageIndices{PECTSP}
}
\author{
\packageAuthor{PECTSP}

Maintainer: \packageMaintainer{PECTSP}
}
\references{
TO BE ADDED LATER
}
\keyword{ package }
\examples{
}
