\name{TimSPEC-package}
\alias{TimSPEC-package}
\alias{TimSPEC}
\docType{package}
\title{\packageTitle{TimSPEC}}
\description{\packageDescription{TimSPEC}}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TimSPEC}
\packageIndices{TimSPEC}
}
\author{
\packageAuthor{TimSPEC}

Maintainer: \packageMaintainer{TimSPEC}
}
\references{
TO BE ADDED LATER
}
\keyword{ package }
